[refactor] Split model tests into package from single file
